backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
authres==1.2.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
fastapi==0.128.0
filelock==3.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipaddress==1.0.23
joblib==1.5.3
mail-parser==4.1.4
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-magic==0.4.27
requests==2.32.5
requests-file==3.0.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
soupsieve==2.8.3
starlette==0.50.0
threadpoolctl==3.6.0
tldextract==5.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
sentence-transformers==3.0.0
